+++ rust-fsm/tests/circuit_breaker.rs
@@ -78,7 +78,7 @@ fn circuit_breaker() {
     // Set up a timer
     let machine_wait = machine.clone();
     std::thread::spawn(move || {
-        std::thread::sleep(Duration::new(5, 0));
+        std::thread::sleep(Duration::from_millis(500));
         let mut lock = machine_wait.lock().unwrap();
         let res = lock.consume(&CircuitBreakerInput::TimerTriggered).unwrap();
         assert_eq!(res, None);
@@ -88,7 +88,7 @@ fn circuit_breaker() {
     // Try to pass a request when the circuit breaker is still open
     let machine_try = machine.clone();
     std::thread::spawn(move || {
-        std::thread::sleep(Duration::new(1, 0));
+        std::thread::sleep(Duration::from_millis(100));
         let mut lock = machine_try.lock().unwrap();
         let res = lock.consume(&CircuitBreakerInput::Successful);
         assert!(matches!(res, Err(TransitionImpossibleError)));
@@ -96,7 +96,7 @@ fn circuit_breaker() {
     });
 
     // Test if the circit breaker was actually closed
-    std::thread::sleep(Duration::new(7, 0));
+    std::thread::sleep(Duration::from_millis(700));
     {
         let mut lock = machine.lock().unwrap();
         let res = lock.consume(&CircuitBreakerInput::Successful).unwrap();